feat: rename Import to ImportSeed
